Clean old code and move helper functions (#8946)
